fix: move widget to "EodashMap.vue"
